+++ src/for_2D_build/common/data_type.h
@@ -30,6 +30,7 @@
 #define DATA_TYPE_2D_H
 
 #include "base_data_type.h"
+#include "geometric_primitive.h"
 #include "scalar_functions.h"
 
 namespace SPH
@@ -37,7 +38,6 @@ namespace SPH
 using Arrayi = Array2i;
 using Vecd = Vec2d;
 using Matd = Mat2d;
-using AlignedBox = AlignedBox2d;
 using AngularVecd = Real;
 using Rotation = Rotation2d;
 using BoundingBox = BaseBoundingBox<Vec2d>;
