+++ src/for_3D_build/common/data_type.h
@@ -29,6 +29,7 @@
 #define DATA_TYPE_3D_H
 
 #include "base_data_type.h"
+#include "geometric_primitive.h"
 #include "scalar_functions.h"
 
 namespace SPH
@@ -36,7 +37,6 @@ namespace SPH
 using Arrayi = Array3i;
 using Vecd = Vec3d;
 using Matd = Mat3d;
-using AlignedBox = AlignedBox3d;
 using AngularVecd = Vec3d;
 using Rotation = Rotation3d;
 using BoundingBox = BaseBoundingBox<Vec3d>;
